\name{sortedNeighbourhood-package} 
\alias{sortedNeighbourhood-package}
\alias{sortedNeighbourhood}
\docType{package}
\title{
Similarity Matching by Sorted Neighbourhood Method
}
\description{
This package contains functions for searching similar objects
in two files of data.

Currently the package expects the data to consist of binary vectors
(fingerprints). The measure of similarity is calculated by the
Tanimoto-Coefficient.

The package's data structure and algorithms are based on
Sorted Neighbourhood and XOR-folding
as discribed in the article referenced below.
}
\details{
\tabular{ll}{
Package: \tab sortedNeighbourhood\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-02-19\cr
License: \tab GPL-3\cr
}
}
\author{
Stefani Kamphausen, Markus Koetter, iQbis consulting GmbH

Maintainer: Prof. Dr. Rainer Schnell <rainer.schnell@uni-due.de>
}
\references{
Hernandez MA, Stolfo SJ:
Real-world Data is Dirty: Data Cleansing and The Merge/Purge Problem
Data Mining and Knowledge Discovery, 1998

Baldi P, Hirschberg DS, Nasr RJ:
Speeding Up Chemical Database Searches
Using a Proximity Filter Based on the Logical Exclusive OR.
Journal of Chemical Information and Modeling 2008
}
\keyword{misc}
